requirements/dev.txt
django-nose==0.1
nose==0.11.3
fabric==0.9.2
pylint==0.21.1
django-extensions==0.5
django-debug-toolbar
Paver==1.0.3
mockito==0.5.1

